requirements.txt
gkeepapi
flox-lib
winotify
